backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tutor"
version = "0.0.0"
description = "Provides functionality for the tutor project"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Natural Language :: English",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy==2.2.4",
    "pandas==2.2.3",
    "matplotlib==3.10.1",
    "requests==2.32.3",
    "python-dotenv==1.1.0",
    "fastparquet==2024.11.0",
    "seaborn==0.13.2",
    "scikit-learn==1.6.1",
    "transformers==4.51.3",
    "torch==2.7.0",
    "nltk==3.9.1",
    "flask==3.1.1",
    "flask-cors==6.0.0",
    "protobuf>=5.29.5",
    "accelerate==1.7.0",
    "peft==0.15.2",
    "sentencepiece==0.2.0",
    "accelerate==1.7.0",
    "google-generativeai==0.8.5",
    "openpyxl==3.1.5"
]
